+++ src/nips/nip25.rs
@@ -34,7 +34,7 @@ impl Client {
     /// let identity = Identity::from_str(env!("SECRET_KEY")).unwrap();
     ///
     /// // Here we react to an event
-    /// client.react_to(&identity, "342060554ca30a9792f6e6959675ae734aed02c23e35037d2a0f72ac6316e83d", "884704bd421721e292edbff42eb77547fe115c6ff9825b08fc366be4cd69e9f6", "+").unwrap();
+    /// client.react_to(&identity, "342060554ca30a9792f6e6959675ae734aed02c23e35037d2a0f72ac6316e83d", "884704bd421721e292edbff42eb77547fe115c6ff9825b08fc366be4cd69e9f6", "+", 0).unwrap();
     /// ```
     pub fn react_to(
         &mut self,
@@ -42,6 +42,7 @@ impl Client {
         event_id: &str,
         event_pub_key: &str,
         reaction: &str,
+        difficulty_target: u16,
     ) -> Result<Event, NIP25Error> {
         let event = EventPrepare {
             pub_key: identity.public_key_str.clone(),
@@ -53,7 +54,7 @@ impl Client {
             ],
             content: reaction.to_string(),
         }
-        .to_event(identity);
+        .to_event(identity, difficulty_target);
 
         self.publish_event(&event)?;
         Ok(event)
@@ -67,15 +68,16 @@ impl Client {
     /// use std::str::FromStr;
     /// let mut client = Client::new(vec![env!("RELAY_URL")]).unwrap();
     /// let identity = Identity::from_str(env!("SECRET_KEY")).unwrap();
-    /// client.like(&identity, "342060554ca30a9792f6e6959675ae734aed02c23e35037d2a0f72ac6316e83d", "884704bd421721e292edbff42eb77547fe115c6ff9825b08fc366be4cd69e9f6").unwrap();
+    /// client.like(&identity, "342060554ca30a9792f6e6959675ae734aed0223e35037d2a0f72ac6316e83d", "884704bd421721e292edbff42eb77547fe115c6ff9825b08fc366be4cd69e9f6", 0).unwrap();
     /// ```
     pub fn like(
         &mut self,
         identity: &Identity,
         event_id: &str,
         event_pub_key: &str,
+        difficulty_target: u16,
     ) -> Result<Event, NIP25Error> {
-        self.react_to(identity, event_id, event_pub_key, "+")
+        self.react_to(identity, event_id, event_pub_key, "+", difficulty_target)
     }
 
     /// Add a dislike to an event
@@ -86,14 +88,15 @@ impl Client {
     /// use std::str::FromStr;
     /// let mut client = Client::new(vec![env!("RELAY_URL")]).unwrap();
     /// let identity = Identity::from_str(env!("SECRET_KEY")).unwrap();
-    /// client.dislike(&identity, "342060554ca30a9792f6e6959675ae734aed02c23e35037d2a0f72ac6316e83d", "884704bd421721e292edbff42eb77547fe115c6ff9825b08fc366be4cd69e9f6").unwrap();
+    /// client.dislike(&identity, "342060554ca30a9792f6e6959675ae734aed02c23e35037d2a0f72ac6316e83d", "884704bd421721e292edbff42eb77547fe115c6ff9825b08fc366be4cd69e9f6", 0).unwrap();
     /// ```
     pub fn dislike(
         &mut self,
         identity: &Identity,
         event_id: &str,
         event_pub_key: &str,
+        difficulty_target: u16,
     ) -> Result<Event, NIP25Error> {
-        self.react_to(identity, event_id, event_pub_key, "-")
+        self.react_to(identity, event_id, event_pub_key, "-", difficulty_target)
     }
 }