+++ src/nips/nip9.rs
@@ -28,18 +28,19 @@ impl Client {
     /// let mut client = Client::new(vec![env!("RELAY_URL")]).unwrap();
     /// let identity = Identity::from_str(env!("SECRET_KEY")).unwrap();
     /// // Create an event
-    /// let event = client.publish_text_note(&identity, "Hello Nostr! :)", &[])
+    /// let event = client.publish_text_note(&identity, "Hello Nostr! :)", &[], 0)
     ///   .unwrap();
     ///
     /// // Delete the event
-    /// client.delete_event(&identity, &event.id).unwrap();
+    /// client.delete_event(&identity, &event.id, 0).unwrap();
     /// ```
     pub fn delete_event(
         &mut self,
         identity: &Identity,
         event_id: &str,
+        difficulty_target: u16,
     ) -> Result<Event, NIP9Error> {
-        self.delete_event_with_reason(identity, event_id, "")
+        self.delete_event_with_reason(identity, event_id, "", difficulty_target)
     }
 
     /// Delete an event with a reason
@@ -51,17 +52,18 @@ impl Client {
     /// let mut client = Client::new(vec![env!("RELAY_URL")]).unwrap();
     /// let identity = Identity::from_str(env!("SECRET_KEY")).unwrap();
     /// // Create an event
-    /// let event = client.publish_text_note(&identity, "Hello Nostr! :)", &[])
+    /// let event = client.publish_text_note(&identity, "Hello Nostr! :)", &[], 0)
     ///  .unwrap();
     ///
     /// // Delete the event with a reason
-    /// client.delete_event_with_reason(&identity, &event.id, "This is a reason").unwrap();
+    /// client.delete_event_with_reason(&identity, &event.id, "This is a reason", 0).unwrap();
     /// ```
     pub fn delete_event_with_reason(
         &mut self,
         identity: &Identity,
         event_id: &str,
         reason: &str,
+        difficulty_target: u16,
     ) -> Result<Event, NIP9Error> {
         let event = EventPrepare {
             pub_key: identity.public_key_str.clone(),
@@ -70,7 +72,7 @@ impl Client {
             tags: vec![vec!["e".to_string(), event_id.to_string()]],
             content: reason.to_string(),
         }
-        .to_event(identity);
+        .to_event(identity, difficulty_target);
 
         self.publish_event(&event)?;
         Ok(event)